+++ NEWS.rst
@@ -1,3 +1,16 @@
+Version 1.10.2, 2020-11-13
+==========================
+
+New features
+------------
+
+* #39244: Added root filter to orgunit search
+* #39244: Unified the paged_filtered_get and paged_get methods, thus changing
+  the order of paged lists throughout MO. The order is now by UUID instead of
+  by user_key / bvn.
+* #39468: Fix bug where UI query page didn't work with Excel files
+
+
 Version 1.10.1, 2020-10-26
 ==========================